requirements.txt
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
idna==3.8
jaconv==0.4.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mutagen==1.47.0
platformdirs==4.2.2
pycryptodomex==3.20.0
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
pykakasi==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytube==15.0.0
rapidfuzz==3.9.7
redis==5.0.8
requests==2.32.3
rich==13.8.0
six==1.16.0
sniffio==1.3.1
soundcloud-v2==1.6.0
soupsieve==2.6
spotdl==4.2.8
spotipy==2.24.0
starlette==0.27.0
syncedlyrics==1.0.1
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.2
uvicorn==0.23.2
websockets==13.0.1
wrapt==1.16.0
yt-dlp==2024.8.6
ytmusicapi==1.8.1
